requirements/local.txt
# The requirement file to be used on local development system
-r base.txt
coverage==3.7
django-debug-toolbar==0.11.0
